fix(sizers): remove the 1x1-fast _dsp exception and the 1xN routing overflow
Two conv-sizer routing holes proven during #364's review:

arm_convolve_1x1_s8_fast_get_buffer_size_dsp's non-armclang branch
returned 0 reading nothing, so arm_convolve_wrapper_s8_get_buffer_size_dsp
answered 0 for 1x1 shapes its dispatcher rejects with -1 -- the
#318/#323/#348 defect class on its last leg, stated as the one
deliberate exception in #364's README and header notes. The gate now
sits ahead of the compiler split in the helper and, per the #364
pattern, ahead of the dispatch in
arm_convolve_1x1_s8_fast_get_buffer_size: -1 for a negative channel
count or an armclang-formula byte count over INT32_MAX, on every
compiler and target, so the three wrapper entry points answer alike
and the README/header "one exception" carve-outs are removed.

arm_nn_is_convolve_1_x_n multiplied stride.w by the channel count in
int32 before any guard -- signed-overflow UB reachable from the public
wrapper sizer (stride.w=65536, c=65536, UBSan-proven). The multiply is
folded to 64 bits; the mod-4 routing answer is unchanged for every
non-overflowing input (1280-probe differential sweep against
origin/main, bit-identical).

Verified on host under -fsanitize=undefined -fno-sanitize-recover=all:
both issue repros fail/trap on origin/main and run clean fixed; a
22-pair probe sweep (negative, zero, valid, both sides of the exact
536870911/536870912 overflow boundary, fast and non-fast 1x1) shows
dispatcher, _dsp and _mve wrappers agreeing on every probe. Two Unity
cases land in the existing 1x1-fast and 1xN suites and fail against
the old code (wrong 0 from the _dsp leg; UBSan trap in the predicate).

Refs #366, #367, #364, #349, #323, #288.

+++ Include/arm_nnsupportfunctions.h
@@ -364,8 +364,11 @@ __STATIC_INLINE bool arm_nn_is_convolve_1_x_n(const cmsis_nn_conv_params *conv_p
                                               const cmsis_nn_dims *input_dims,
                                               const cmsis_nn_dims *filter_dims)
 {
+    // stride.w and c are each in range on their own, so their product can exceed INT32_MAX; the multiply is
+    // folded to 64 bits because a wrapped int32 product is signed-overflow UB, reachable from the public conv
+    // sizers (issue #367). The remainder test is unchanged for any product that fits in an int32_t.
     return (input_dims->h == 1) && (conv_params->dilation.w == 1) && (filter_dims->h == 1) &&
-        ((conv_params->stride.w * input_dims->c) % 4 == 0) && (input_dims->c == filter_dims->c);
+        (((int64_t)conv_params->stride.w * (int64_t)input_dims->c) % 4 == 0) && (input_dims->c == filter_dims->c);
 }
 
 /**
